#1209 move `failOnError` and `failOnWarning` completely into `OptimizeMojo` from `TranspileMojo`
