+++ eo-maven-plugin/src/test/java/org/eolang/maven/TranspileMojoTest.java
@@ -24,14 +24,6 @@
 package org.eolang.maven;
 
 import com.jcabi.log.Logger;
-import com.jcabi.xml.XML;
-import com.jcabi.xml.XMLDocument;
-import com.jcabi.xml.XSLDocument;
-import com.yegor256.xsline.Shift;
-import com.yegor256.xsline.StXSL;
-import com.yegor256.xsline.TrDefault;
-import com.yegor256.xsline.Xsline;
-import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import org.apache.maven.plugin.testing.stubs.MavenProjectStub;
@@ -228,112 +220,6 @@ final class TranspileMojoTest {
     }
 
     @Test
-    void testFailOnWarning(@TempDir final Path temp) throws Exception {
-        final Path src = temp.resolve("foo.src.eo");
-        new Home().save(new ResourceOf("org/eolang/maven/withwarning.eo"), src);
-        final Path target = temp.resolve("target");
-        final Path generated = temp.resolve("generated");
-        final Path foreign = temp.resolve("eo-foreign.json");
-        final Path transpiled = temp.resolve("eo-transpiled");
-        Catalogs.INSTANCE.make(foreign)
-            .add("foo.src")
-            .set(AssembleMojo.ATTR_SCOPE, "compile")
-            .set(AssembleMojo.ATTR_EO, src.toString());
-        new Moja<>(ParseMojo.class)
-            .with("targetDir", target.toFile())
-            .with("foreign", foreign.toFile())
-            .with("cache", temp.resolve("cache/parsed"))
-            .with("foreignFormat", "csv")
-            .execute();
-        new Moja<>(OptimizeMojo.class)
-            .with("targetDir", target.toFile())
-            .with("foreign", foreign.toFile())
-            .with("foreignFormat", "csv")
-            .with("failOnError", false)
-            .execute();
-        this.applyXsl(
-            "org/eolang/maven/set-warning-severity.xsl",
-            target.resolve("03-optimize/foo/src.xmir")
-        );
-        final IllegalStateException exception = Assertions.assertThrows(
-            IllegalStateException.class,
-            () -> new Moja<>(TranspileMojo.class)
-                .with("project", new MavenProjectStub())
-                .with("targetDir", target.toFile())
-                .with("generatedDir", generated.toFile())
-                .with("foreign", foreign.toFile())
-                .with("foreignFormat", "csv")
-                .with("failOnWarning", true)
-                .with("transpiled", transpiled.toFile())
-                .with("transpiledFormat", "csv")
-                .execute()
-        );
-        MatcherAssert.assertThat(
-            exception.getMessage(),
-            Matchers.equalTo("There are 1 warning(s) in foo.src, see log above")
-        );
-    }
-
-    @Test
-    void testFailOnError(@TempDir final Path temp) throws Exception {
-        final Path wrong = temp.resolve("foo.wrong.eo");
-        final Path right = temp.resolve("foo.right.eo");
-        new Home().save(new ResourceOf("org/eolang/maven/witherror.eo"), wrong);
-        new Home().save(new ResourceOf("org/eolang/maven/mess.eo"), right);
-        final Path target = temp.resolve("target");
-        final Path generated = temp.resolve("generated");
-        final Path foreign = temp.resolve("eo-foreign.json");
-        final Path transpiled = temp.resolve("eo-transpiled");
-        Catalogs.INSTANCE.make(foreign)
-            .add("foo.wrong")
-            .set(AssembleMojo.ATTR_SCOPE, "compile")
-            .set(AssembleMojo.ATTR_EO, wrong.toString());
-        Catalogs.INSTANCE.make(foreign)
-            .add("foo.right")
-            .set(AssembleMojo.ATTR_SCOPE, "compile")
-            .set(AssembleMojo.ATTR_EO, right.toString());
-        new Moja<>(ParseMojo.class)
-            .with("targetDir", target.toFile())
-            .with("foreign", foreign.toFile())
-            .with("cache", temp.resolve("cache/parsed"))
-            .with("foreignFormat", "csv")
-            .with("failOnError", false)
-            .execute();
-        new Moja<>(OptimizeMojo.class)
-            .with("targetDir", target.toFile())
-            .with("foreign", foreign.toFile())
-            .with("foreignFormat", "csv")
-            .execute();
-        new Moja<>(TranspileMojo.class)
-            .with("project", new MavenProjectStub())
-            .with("targetDir", target.toFile())
-            .with("generatedDir", generated.toFile())
-            .with("foreign", foreign.toFile())
-            .with("foreignFormat", "csv")
-            .with("failOnError", false)
-            .with("transpiled", transpiled.toFile())
-            .with("transpiledFormat", "csv")
-            .execute();
-        final Path mess = generated.resolve("EOorg/EOeolang/EOexamples/EOmessTest.java");
-        final Path main = generated.resolve("EOorg/EOeolang/EOexamples/EOmainTest.java");
-        MatcherAssert.assertThat(
-            String.format("The file \"%s\" wasn't created", mess),
-            new Home().exists(mess),
-            Matchers.is(true)
-        );
-        MatcherAssert.assertThat(
-            String.format("The file \"%s\" was created", main),
-            Files.notExists(main),
-            Matchers.is(true)
-        );
-        Assertions.assertTrue(mess.toFile().setLastModified(0L));
-        final Path xmir = target.resolve("06-transpile")
-            .resolve("foo")
-            .resolve("right.xmir");
-        Assertions.assertTrue(new Home().exists(xmir));
-    }
-
-    @Test
     void testSimpleCompilation(@TempDir final Path temp)
         throws Exception {
         final String java = this.compile(
@@ -428,22 +314,5 @@ final class TranspileMojoTest {
         final String out = new TextOf(new InputOf(java)).asString();
         Logger.debug(this, "Java output:\n%s", out);
         return out;
-    }
-
-    /**
-     * Apply XSL transformation.
-     * @param xsl Path to XSL within classpath
-     * @param xml Path to XML to be tranformed
-     */
-    private void applyXsl(final String xsl, final Path xml) throws Exception {
-        final XML output = new Xsline(
-            new TrDefault<Shift>()
-                .with(
-                    new StXSL(
-                        new XSLDocument(
-                            new ResourceOf(xsl).stream()
-                )))
-        ).pass(new XMLDocument(xml));
-        new Home().save(output.toString(), xml);
     }
 }
