+++ eo-parser/src/main/resources/org/eolang/parser/errors/fail-on-warnings.xsl
@@ -23,14 +23,14 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 -->
 <!--
-Raise an error if errors are found within program
+Raise an error if warnings are found within program
 -->
-<xsl:stylesheet xmlns:xsl="http://www.w3.org/1999/XSL/Transform" xmlns:eo="https://www.eolang.org" xmlns:xs="http://www.w3.org/2001/XMLSchema" id="fail-on-errors" version="2.0">
+<xsl:stylesheet xmlns:xsl="http://www.w3.org/1999/XSL/Transform" xmlns:eo="https://www.eolang.org" xmlns:xs="http://www.w3.org/2001/XMLSchema" id="fail-on-warnings" version="2.0">
   <xsl:output encoding="UTF-8" method="xml"/>
-  <xsl:template match="/program/errors/error[@severity='error']">
+  <xsl:template match="/program/errors/error[@severity='warning']">
     <xsl:message terminate="yes">
-      <xsl:text>Errors identified: </xsl:text>
-      <xsl:for-each select="/program/errors/error[@severity='error']">
+      <xsl:text>Warnings identified: </xsl:text>
+      <xsl:for-each select="/program/errors/error[@severity='warning']">
         <xsl:value-of select="concat(text(), '; ')"/>
       </xsl:for-each>
     </xsl:message>
